adds edit button to show modal that opens edit modal and sents put request to update reminder action
